agent/requirements.txt
livekit-agents>=1.0.23
livekit-plugins-openai>=1.0.23
livekit-plugins-deepgram>=1.0.23
livekit-plugins-silero>=1.0.23
livekit-protocol~=1.0.3
python-dotenv~=1.0
tiktoken~=0.8.0
langchain~=0.3.7
langchain_openai
openai~=1.82.1
supabase~=2.10.0
fastapi~=0.115.6
uvicorn~=0.32.1
aiohttp~=3.11.11
twilio~=9.4.5
aiosmtplib~=4.0.0
aiofiles~=24.1.0
tenacity~=9.0.0